refactor: extract AddExistingExpenseTable to match Tables/ convention
