+++ src/pages/AddExistingExpense.tsx
@@ -4,10 +4,9 @@ import HeaderWithBackButton from '@components/HeaderWithBackButton';
 import ScreenWrapper from '@components/ScreenWrapper';
 import ScrollView from '@components/ScrollView';
 import UnreportedExpensesSkeleton from '@components/Skeletons/UnreportedExpensesSkeleton';
-import type {CompareItemsCallback, FilterConfig, IsItemInFilterCallback, IsItemInSearchCallback, TableColumn, TableData} from '@components/Table';
-import Table from '@components/Table';
+import type {UnreportedExpenseTableRowData} from '@components/Tables/AddExistingExpenseTable';
+import AddExistingExpenseTable from '@components/Tables/AddExistingExpenseTable';
 
-import {useCurrencyListActions} from '@hooks/useCurrencyList';
 import useCurrentUserPersonalDetails from '@hooks/useCurrentUserPersonalDetails';
 import useIsInLandscapeMode from '@hooks/useIsInLandscapeMode';
 import {useMemoizedLazyIllustrations} from '@hooks/useLazyAsset';
@@ -22,8 +21,7 @@ import getNonEmptyStringOnyxID from '@libs/getNonEmptyStringOnyxID';
 import interceptAnonymousUser from '@libs/interceptAnonymousUser';
 import type {AddExistingExpensesParamList} from '@libs/Navigation/types';
 import {shouldRestrictUserBillableActions} from '@libs/SubscriptionUtils';
-import tokenizedSearch from '@libs/tokenizedSearch';
-import {createUnreportedExpenses, getAmount, getCreated, getCurrency, getDescription, getEligibleTransactionsToAdd, getMerchant, isUnreportedTransaction} from '@libs/TransactionUtils';
+import {createUnreportedExpenses, getEligibleTransactionsToAdd} from '@libs/TransactionUtils';
 
 import Navigation from '@navigation/Navigation';
 import type {PlatformStackScreenProps} from '@navigation/PlatformStackNavigation/types';
@@ -39,21 +37,14 @@ import {validTransactionDraftIDsSelector} from '@src/selectors/TransactionDraft'
 import type Transaction from '@src/types/onyx/Transaction';
 import getEmptyArray from '@src/types/utils/getEmptyArray';
 
-import type {ListRenderItemInfo} from '@shopify/flash-list';
 import type {OnyxCollection} from 'react-native-onyx';
 
 import {useCallback, useEffect, useState} from 'react';
 import {View} from 'react-native';
 
-import AddExistingExpenseTableRow from './AddExistingExpenseTableRow';
-
 type AddExistingExpensePageType = PlatformStackScreenProps<AddExistingExpensesParamList, typeof SCREENS.ADD_EXISTING_EXPENSES_ROOT>;
-type ExpenseTableColumnKey = 'date' | 'amount';
-type ExpenseTableFilterKey = 'status';
-type UnreportedExpenseTableRowData = Transaction & TableData;
 
 function AddExistingExpense({route}: AddExistingExpensePageType) {
-    const {convertToDisplayString} = useCurrencyListActions();
     const {translate} = useLocalize();
     const illustrations = useMemoizedLazyIllustrations(['FolderWithPapersAndWatch']);
     const [errorMessage, setErrorMessage] = useState<string>('');
@@ -121,86 +112,6 @@ function AddExistingExpense({route}: AddExistingExpensePageType) {
             setErrorMessage('');
         }
     };
-
-    const columns: Array<TableColumn<ExpenseTableColumnKey>> = [
-        {key: 'date', label: translate('common.date'), sortable: true},
-        {key: 'amount', label: translate('iou.amount'), sortable: true},
-    ];
-
-    const filters: FilterConfig<ExpenseTableFilterKey> = {
-        status: {
-            label: translate('common.status'),
-            filterType: CONST.TABLES.FILTER_TYPE.MULTI_SELECT,
-            options: [
-                {label: translate('common.unreported'), value: CONST.SEARCH.STATUS.EXPENSE.UNREPORTED},
-                {label: translate('common.draft'), value: CONST.SEARCH.STATUS.EXPENSE.DRAFTS},
-            ],
-        },
-    };
-
-    const isItemInFilter: IsItemInFilterCallback<UnreportedExpenseTableRowData> = (item, values) => {
-        if (values.length === 0) {
-            return true;
-        }
-
-        const includesUnreported = values.includes(CONST.SEARCH.STATUS.EXPENSE.UNREPORTED);
-        const includesDrafts = values.includes(CONST.SEARCH.STATUS.EXPENSE.DRAFTS);
-        const isUnreported = isUnreportedTransaction(item);
-
-        if (includesUnreported && isUnreported) {
-            return true;
-        }
-        return includesDrafts && !isUnreported;
-    };
-
-    const isItemInSearch: IsItemInSearchCallback<UnreportedExpenseTableRowData> = (item, searchValue) => {
-        const results = tokenizedSearch([item], searchValue, (transaction) => {
-            const searchableFields: string[] = [];
-
-            const merchant = getMerchant(transaction);
-            if (merchant !== CONST.TRANSACTION.PARTIAL_TRANSACTION_MERCHANT && merchant !== CONST.TRANSACTION.DEFAULT_MERCHANT) {
-                searchableFields.push(merchant);
-            }
-
-            const description = getDescription(transaction);
-            if (description.trim()) {
-                searchableFields.push(description);
-            }
-
-            const amount = getAmount(transaction);
-            const currency = getCurrency(transaction);
-            const formattedAmount = convertToDisplayString(amount, currency);
-            searchableFields.push(formattedAmount);
-
-            // This allows users to search "2000" and find "$2,000.00" for example
-            const normalizedAmount = (amount / 100).toString();
-            searchableFields.push(normalizedAmount);
-
-            return searchableFields;
-        });
-
-        return results.length > 0;
-    };
-
-    const compareItems: CompareItemsCallback<UnreportedExpenseTableRowData, ExpenseTableColumnKey> = (item1, item2, activeSorting) => {
-        const orderMultiplier = activeSorting.order === 'desc' ? -1 : 1;
-
-        if (activeSorting.columnKey === 'amount') {
-            return (getAmount(item1) - getAmount(item2)) * orderMultiplier;
-        }
-
-        // Default: sort by the expense date, matching how Spend > Expenses orders rows by default.
-        const created1 = new Date(getCreated(item1)).getTime();
-        const created2 = new Date(getCreated(item2)).getTime();
-        return (created1 - created2) * orderMultiplier;
-    };
-
-    const renderItem = ({item, index}: ListRenderItemInfo<UnreportedExpenseTableRowData>) => (
-        <AddExistingExpenseTableRow
-            item={item}
-            rowIndex={index}
-        />
-    );
 
     const paginationFooterContent = shouldShowUnreportedTransactionsSkeletons ? <UnreportedExpensesSkeleton fixedNumberOfItems={3} /> : undefined;
 
@@ -287,34 +198,14 @@ function AddExistingExpense({route}: AddExistingExpensePageType) {
                 onBackButtonPress={Navigation.goBack}
             />
             <View style={styles.flex1}>
-                <Table<UnreportedExpenseTableRowData, ExpenseTableColumnKey, ExpenseTableFilterKey>
+                <AddExistingExpenseTable
                     data={unreportedExpenses}
-                    title={translate('common.expenses')}
-                    columns={columns}
-                    selectionEnabled
-                    shouldEnableSelectionInNarrowPaneModal
                     selectedKeys={[...selectedIds]}
                     onRowSelectionChange={onRowSelectionChange}
-                    initialSortColumn="date"
-                    initialSortOrder="desc"
-                    compareItems={compareItems}
-                    filters={filters}
-                    isItemInFilter={isItemInFilter}
-                    isItemInSearch={isItemInSearch}
-                    renderItem={renderItem}
-                    keyExtractor={(item) => item.keyForList}
                     onEndReached={fetchMoreUnreportedTransactions}
                     onEndReachedThreshold={0.75}
-                    maintainVisibleContentPosition={{disabled: true}}
                     ListFooterComponent={paginationFooterContent}
-                >
-                    <Table.ListHeader>
-                        <Table.FilterBar label={translate('iou.findExpense')} />
-                    </Table.ListHeader>
-                    <Table.NoResultsState />
-                    <Table.Header />
-                    <Table.Body />
-                </Table>
+                />
             </View>
             {footerContent}
         </ScreenWrapper>
@@ -322,4 +213,3 @@ function AddExistingExpense({route}: AddExistingExpensePageType) {
 }
 
 export default AddExistingExpense;
-export type {UnreportedExpenseTableRowData};
